# risk_brain_reporter/openapi.yaml
openapi: 3.0.3
info:
  title: Risk Brain Reporter API
  description: |
    Deterministic governance artefact generator for Risk Brain shadow intelligence.
    
    **Service Responsibilities (Locked):**
    - ✅ Pull metrics
    - ✅ Aggregate into canonical snapshot
    - ✅ Render PDFs
    - ✅ Write immutable artefacts
    - ✅ Emit telemetry
    
    **Never Allowed:**
    - ❌ Emit commands
    - ❌ Touch ledgers
    - ❌ Influence runtime policy
    - ❌ Call Kafka producers
    - ❌ Interact with A-domain APIs
  version: 1.0.0
  contact:
    name: TuringCore National Infrastructure Team
    email: risk-brain@turingcore.com

servers:
  - url: https://risk-brain-reporter.internal.turingcore.com/api/v1
    description: Internal production server
  - url: http://localhost:8080/api/v1
    description: Local development server

tags:
  - name: Weekly Reports
    description: Weekly board pack generation
  - name: Regulator Annexes
    description: On-demand regulator annex generation
  - name: Audit
    description: Report listing and audit trail

paths:
  /reports/weekly/run:
    post:
      tags:
        - Weekly Reports
      summary: Run weekly board pack generation (Network + All Tenants)
      description: |
        Scheduled cron only. Generates PDFs for every tenant + network pack.
      operationId: runWeeklyBoardPackAll
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - week
              properties:
                week:
                  type: string
                  pattern: '^\d{4}-W\d{2}$'
                  example: '2025-W49'
                  description: ISO week number (YYYY-WXX)
                force_regenerate:
                  type: boolean
                  default: false
                  description: Force regeneration even if report already exists
      responses:
        '200':
          description: Weekly board pack generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyRunResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

  /reports/weekly/run/{tenant_id}:
    post:
      tags:
        - Weekly Reports
      summary: Run weekly board pack generation (Single Tenant)
      description: |
        Generate weekly board pack for a single tenant.
      operationId: runWeeklyBoardPackTenant
      security:
        - ServiceToken: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
          example: cu_123
          description: Tenant ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - week
              properties:
                week:
                  type: string
                  pattern: '^\d{4}-W\d{2}$'
                  example: '2025-W49'
                  description: ISO week number (YYYY-WXX)
                force_regenerate:
                  type: boolean
                  default: false
                  description: Force regeneration even if report already exists
      responses:
        '200':
          description: Weekly board pack generation started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyRunTenantResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Tenant not found
        '500':
          description: Internal server error

  /reports/weekly/status/{tenant_id}/{week}:
    get:
      tags:
        - Weekly Reports
      summary: Get weekly board pack status
      description: |
        Get the status of a weekly board pack generation.
      operationId: getWeeklyBoardPackStatus
      security:
        - ServiceToken: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
          example: cu_123
          description: Tenant ID
        - name: week
          in: path
          required: true
          schema:
            type: string
            pattern: '^\d{4}-W\d{2}$'
          example: '2025-W49'
          description: ISO week number (YYYY-WXX)
      responses:
        '200':
          description: Weekly board pack status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WeeklyStatusResponse'
        '401':
          description: Unauthorized
        '404':
          description: Report not found
        '500':
          description: Internal server error

  /reports/regulator/run/{tenant_id}:
    post:
      tags:
        - Regulator Annexes
      summary: Generate regulator annex (On Demand)
      description: |
        Generate on-demand regulator annex for a specific period.
      operationId: generateRegulatorAnnex
      security:
        - ServiceToken: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
          example: cu_123
          description: Tenant ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - period_start
                - period_end
              properties:
                period_start:
                  type: string
                  format: date
                  example: '2025-12-01'
                  description: Period start date (YYYY-MM-DD)
                period_end:
                  type: string
                  format: date
                  example: '2025-12-07'
                  description: Period end date (YYYY-MM-DD)
                include_event_samples:
                  type: boolean
                  default: true
                  description: Include anonymised event samples
      responses:
        '200':
          description: Regulator annex generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegulatorAnnexResponse'
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '404':
          description: Tenant not found
        '500':
          description: Internal server error

  /reports/{tenant_id}:
    get:
      tags:
        - Audit
      summary: List reports (Audit)
      description: |
        List all reports for a tenant (weekly board packs + regulator annexes).
      operationId: listReports
      security:
        - ServiceToken: []
      parameters:
        - name: tenant_id
          in: path
          required: true
          schema:
            type: string
          example: cu_123
          description: Tenant ID
      responses:
        '200':
          description: List of reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportListResponse'
        '401':
          description: Unauthorized
        '404':
          description: Tenant not found
        '500':
          description: Internal server error

  /reports/alerts/internal:
    post:
      tags:
        - Audit
      summary: Emit internal alert
      description: |
        Automatically emitted on:
        - Missing metrics
        - Safety violations
        - Storage lock failure
        - Template render crash
      operationId: emitInternalAlert
      security:
        - ServiceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InternalAlert'
      responses:
        '200':
          description: Alert emitted
        '400':
          description: Invalid request
        '401':
          description: Unauthorized
        '500':
          description: Internal server error

components:
  securitySchemes:
    ServiceToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Internal service token (mTLS + static token)

  schemas:
    WeeklyRunResponse:
      type: object
      required:
        - status
        - week
        - tenants_queued
        - started_at
      properties:
        status:
          type: string
          enum: [STARTED]
          example: STARTED
        week:
          type: string
          example: '2025-W49'
        tenants_queued:
          type: integer
          example: 47
          description: Number of tenants queued for report generation
        started_at:
          type: string
          format: date-time
          example: '2025-12-07T23:00:00Z'

    WeeklyRunTenantResponse:
      type: object
      required:
        - status
        - tenant_id
        - week
      properties:
        status:
          type: string
          enum: [STARTED]
          example: STARTED
        tenant_id:
          type: string
          example: cu_123
        week:
          type: string
          example: '2025-W49'

    WeeklyStatusResponse:
      type: object
      required:
        - tenant_id
        - week
        - state
      properties:
        tenant_id:
          type: string
          example: cu_123
        week:
          type: string
          example: '2025-W49'
        state:
          type: string
          enum: [SNAPSHOT_CREATED, PDF_RENDERED, STORED_IMMUTABLY, FAILED]
          example: STORED_IMMUTABLY
        object_path:
          type: string
          example: 's3://risk-brain/weekly/cu_123/risk-brain-week-2025-W49.pdf'
        sha256:
          type: string
          example: 'a92bc13f...'
        generated_at:
          type: string
          format: date-time
          example: '2025-12-07T23:03:21Z'

    RegulatorAnnexResponse:
      type: object
      required:
        - status
        - tenant_id
        - object_path
        - sha256
      properties:
        status:
          type: string
          enum: [STORED_IMMUTABLY]
          example: STORED_IMMUTABLY
        tenant_id:
          type: string
          example: cu_123
        object_path:
          type: string
          example: 's3://risk-brain/regulator/cu_123/annex-2025-12-07.pdf'
        sha256:
          type: string
          example: 'bc28d91f...'

    ReportListResponse:
      type: object
      required:
        - tenant_id
        - weekly_reports
        - regulator_annexes
      properties:
        tenant_id:
          type: string
          example: cu_123
        weekly_reports:
          type: array
          items:
            type: object
            required:
              - week
              - path
              - sha256
            properties:
              week:
                type: string
                example: '2025-W47'
              path:
                type: string
                example: 's3://risk-brain/weekly/cu_123/risk-brain-week-2025-W47.pdf'
              sha256:
                type: string
                example: 'a92bc13f...'
        regulator_annexes:
          type: array
          items:
            type: object
            required:
              - period_end
              - path
              - sha256
            properties:
              period_end:
                type: string
                format: date
                example: '2025-12-07'
              path:
                type: string
                example: 's3://risk-brain/regulator/cu_123/annex-2025-12-07.pdf'
              sha256:
                type: string
                example: 'bc28d91f...'

    InternalAlert:
      type: object
      required:
        - severity
        - tenant_id
        - reason
        - week
      properties:
        severity:
          type: string
          enum: [P1, P2]
          example: P1
        tenant_id:
          type: string
          example: cu_123
        reason:
          type: string
          example: AI_ORIGIN_VIOLATION_NONZERO
        week:
          type: string
          example: '2025-W49'